# api/openapi/auth.yml
openapi: 3.0.1
info:
  title: Mainflux authentication service
  description: HTTP API for managing platform API keys.
  version: "1.0.0"
paths:
  /keys:
    post:
      summary: Issue API key
      description: |
        Generates a new API key. Thew new API key will
        be uniquely identified by its ID.
      tags:
        - auth
      requestBody:
        $ref: "#/components/requestBodies/KeyRequest"
      responses:
        '201':
          description: Issued new key.
        '400':
          description: Failed due to malformed JSON.
        '409':
          description: Failed due to using already existing ID.
        '415':
          description: Missing or invalid content type.
        '500':
          $ref: "#/components/responses/ServiceError"
  /keys/{keyId}:
    get:
      summary: Gets API key details.
      description: |
        Gets API key details for the given key.
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/ApiKeyId"
      responses:
        '200':
          $ref: "#/components/responses/KeyRes"
        '400':
          description: Failed due to malformed query parameters.
        '401':
          description: Missing or invalid access token provided.
        '500':
          $ref: "#/components/responses/ServiceError"
    delete:
      summary: Revoke API key
      description: |
        Revoke API key identified by the given ID.
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/ApiKeyId"
      responses:
        '204':
          description: Key revoked.
        '401':
          description: Missing or invalid access token provided.
        '500':
          $ref: "#/components/responses/ServiceError"
...
  /groups/{groupId}:
    get:
      summary: Gets group info.
      description: |
        Gets info on a group specified by id.
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        '200':
          $ref: "#/components/responses/GroupRes"
        '400':
          description: Failed due to malformed query parameters.
        '401':
          description: Missing or invalid access token provided.
        '404':
          description: Group does not exist.
        '500':
          $ref: "#/components/responses/ServiceError"
    put:
      summary: Updates group data.
      description: |
        Updates Name, Description or Metadata of a group.
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        $ref: "#/components/requestBodies/GroupUpdateReq"
      responses:
        '200':
          description: Group updated.
        '400':
          description: Failed due to malformed query parameters.
        '401':
          description: Missing or invalid access token provided.
        '404':
          description: Group does not exist.
        '500':
          $ref: "#/components/responses/ServiceError"
    delete:
      summary: Deletes group.
      description: |
        Deletes group. If group is parent and descendant groups do not have any members
        child groups will be deleted. Group cannot be deleted if has members or if
        any descendant group has members.
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - $ref: "#/components/parameters/Level"
        - $ref: "#/components/parameters/Metadata"
        - $ref: "#/components/parameters/Tree"
      responses:
        '204':
          description: Group removed.
        '400':
          description: Failed due to malformed query parameters.
        '401':
          description: Missing or invalid access token provided.
        '404':
          description: Group does not exist.
        '500':
          $ref: "#/components/responses/ServiceError"
...
components:
  schemas:
    Key:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "c5747f2f-2a7c-4fe1-b41a-51a5ae290945"
          description: API key unique identifier
        issuer_id:
          type: string
          format: uuid
          example: "9118de62-c680-46b7-ad0a-21748a52833a"
          description: In ID of the entity that issued the token.
        type:
          type: integer
          example: 0
          description: API key type. Keys of different type are processed differently.
        subject:
          type: string
          format: string
          example: "test@example.com"
          description: User's email or service identifier of API key subject.
        issued_at:
          type: string
          format: date-time
          example: "2019-11-26 13:31:52"
          description: Time when the key is generated.
        expires_at:
          type: string
          format: date-time
          example: "2019-11-26 13:31:52"
          description: Time when the Key expires. If this field is missing,
            that means that Key is valid indefinitely.
...
    GroupUpdateSchema:
      type: object
      properties:
        name:
          type: string
          description: |
            Free-form group name. Group name is unique on the given hierarchy level.
        description:
          type: string
          description: Group description, free form text.
        metadata:
          type: object
          description: Arbitrary, object-encoded group's data.
    GroupResSchema:
      type: object
      properties:
        id:
          type: string
          format: ulid
          description: Unique group identifier generated by the service.
        name:
          type: string
          description: Free-form group name.
        parent_id:
          type: string
          description: Group ID of parent group.
        owner_id:
          type: string
          format: uuid
          description: UUID of user that created the group.
        metadata:
          type: object
          description: Arbitrary, object-encoded group's data.
        level:
          type: integer
          description: Level in hierarchy, distance from the root group.
        path:
          type: string
          description: Hierarchy path, concatenated ids of group ancestors.
        children:
          type: object
          # schema: GroupResSchema
        created_at:
          type: string
          description: Datetime of group creation.
        updated_at:
          type: string
          description: Datetime of last group updated.
      required:
        - id
        - name
        - owner_id
        - description
        - level
        - path
        - created_at
        - updated_at
...
  parameters:
    ApiKeyId:
      name: keyId
      description: API Key ID.
      in: path
      schema:
        type: string
        format: uuid
      required: true
...
    GroupId:
      name: groupId
      description: Group ID.
      in: path
      schema:
        type: string
        format: uuid
      required: true
...
    Level:
      name: level
      description: Level of hierarchy up to which to retrieve groups from given group id.
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 5
      required: false
    Metadata:
      name: metadata
      description: Metadata filter. Filtering is performed matching the parameter with metadata on top level. Parameter is json.
      in: query
      required: false
      schema:
        type: object
        additionalProperties: {}
    Tree:
      name: tree
      description: Specify type of response, JSON array or tree.
      in: query
      required: false
      schema:
        type: boolean
        default: false
  requestBodies:
    KeyRequest:
      description: JSON-formatted document describing key request.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              type:
                type: integer
                example: 0
                description: API key type. Keys of different type are processed differently.
              duration:
                type: number
                format: integer
                example: 23456
                description: Number of seconds issued token is valid for.
...
    GroupUpdateReq:
      description: JSON-formatted document describing group create request.
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GroupUpdateSchema"
...
  responses:
    ServiceError:
      description: Unexpected server-side error occurred.
    KeyRes:
      description: Data retrieved.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Key"
...
    GroupRes:
      description: Data retrieved.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GroupResSchema"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        * Users access: "Authorization: Bearer <user_token>"
security:
  - bearerAuth: []